Moved logic from main.rs to lib.rs
